SecuringTheCloud/SecuringTheCloud/requirements.txt
django >= 2.1.6, < 3.0 
pytz
PyDrive
easygui
pycryptodome